problem in the first generated square and move down not smooth
The first square is generated in y=1, not in y=0. But the other squares after this are generated in y=0. There is a gap in moveDown. It moves it until coordinate of y-1 of the where I want to put it then it move down by 1 unit. for example i want to put at coordinate y=6. The moveDown moves it coordinate y=5 then the square moves down by 1 unit. It doesn't do it in one smooth motion.